+++ changelog.md
@@ -10,6 +10,7 @@ If you can't access the Google doc, here is a [PDF](https://github.com/crnormand
 - Kirlian Silvestre/3darkman, fixed Item icon hover picture
 - Added /jb2a command
 - Fixed NaN skill levels for containers
+- Promoted tabbed view
 
 Release 0.10.5 - 6/12/2021